integrated edits from newtoto
