+++ include/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
+/*   By: jdagoy <jdagoy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 16:43:09 by tbarbe            #+#    #+#             */
-/*   Updated: 2023/11/23 08:57:15 by jdagoy           ###   ########.fr       */
+/*   Updated: 2023/11/26 14:59:01 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,12 +51,11 @@ void	prompt_interrupt(void);
 int		signal_set(int sig, void* f);
 
 // redirection
-int redirect(char*** tab);
+int redirect(t_simple_cmd* cmd);
 int read_here_doc(char* limiter);
-int open_file_dup2(char* path, char mode);
+int open_file(char* path, char mode);
 
 //utils.c
-bool* is_piped(void);
 bool* is_piped(void);
 void	set_is_piped(bool value);
 void	print_error(char* str1, char* str2);
